move spawn_named into trace.rs through TraceExt trait
